Introduce deployment options
Deployment options can be set with:
  kubectl gadget deploy --opts <comma-separated-options>

This will set the environment variable $INSPEKTOR_GADGET_OPTIONS in the
gadget DaemonSet. The script entrypoint.sh will then act on it. This
allows to enable or disable some experimental options without
recompiling.

This patch introduces two options:

1. runc_hooks_ldpreload
    Inspektor Gadget currently uses OCI hooks via a patched runc that
    adds PreStart and PostStop hooks. As an alternative option, this
    patch adds support for unpatched runc using a ld.so.preload
    configuration.

2. notraceloop
    Don't start the traceloop daemon. This is useful to avoid the
    memory consumption of the traceloop gadget, if only the other
    gadgets are desired.

+++ gadget-container/entrypoint.sh
@@ -63,5 +63,23 @@ if [ "$FLATCAR_EDGE" = 1 ] ; then
   /bin/gadgettracermanager -serve &
 fi
 
-rm -f /run/traceloop.socket
-exec /bin/traceloop $ARGS
+if [[ "$INSPEKTOR_GADGET_OPTIONS" =~ (^|,)runc_hooks_ldpreload(,|$) ]] ; then
+  echo "Installing ld.so.preload with runchooks.so for OCI hooks"
+  mkdir -p /host/opt/runchooks/
+  cp /opt/runchooks/runchooks.so /host/opt/runchooks/
+  cp /opt/runchooks/add-hooks.jq /host/opt/runchooks/
+  touch /host/etc/ld.so.preload
+  if grep -q ^/opt/runchooks/runchooks.so$ /host/etc/ld.so.preload > /dev/null ; then
+    echo "runchooks.so already setup in /etc/ld.so.preload"
+  else
+    echo "/opt/runchooks/runchooks.so" >> /host/etc/ld.so.preload
+  fi
+fi
+
+if [[ ! "$INSPEKTOR_GADGET_OPTIONS" =~ (^|,)notraceloop(,|$) ]] ; then
+  rm -f /run/traceloop.socket
+  exec /bin/traceloop $ARGS
+fi
+
+echo "Ready."
+sleep infinity
